requirements.txt
ultralytics>=8.3.0
sahi>=0.11.19
supervision>=0.25.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
gradio>=5.0.0
PyYAML>=6.0.2
loguru>=0.7.3
Pillow>=11.0.0
numpy>=1.26.0
matplotlib>=3.10.0
anomalib>=2.1.0
torch>=2.1.0
torchvision>=0.16.0
lightning>=2.1.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scipy>=1.11.0
